Finish last example. Improve formatting. Fix typos.
Complete the last example in step 6 of using the agent to make code
changes to the agent. Also, add example of using the modified agent
(with the colorization changes shown).

Change all the `shell` code blocks to `text` blocks and do manual
formatting of the shell parts.

Fix some typos that introduced an extraneous double-quote.